BE/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
ddgs==9.9.3
distro==1.9.0
dotenv==0.9.9
exa-py==1.16.1
fake-useragent==2.2.0
fastapi==0.123.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
gritql==0.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.0
langchain-cli==0.0.37
langchain-community==0.4.1
langchain-core==1.1.0
langchain-exa==1.0.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langserve==0.3.3
langsmith==0.4.53
lxml==6.0.2
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
primp==0.15.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
shapely==2.1.2
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.44
sse-starlette==1.8.2
starlette==0.50.0
tavily==1.1.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.23.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
psycopg2-binary==2.9.11 
sqlmodel==0.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==42.0.5
